+++ src/main/java/duke/Storage.java
@@ -1,4 +1,4 @@
-package Duke;
+package duke;
 
 import java.io.*;
