+++ src/main/java/duke/Ui.java
@@ -1,6 +1,6 @@
-package Duke;
+package duke;
 
-import Duke.Tasks.Task;
+import duke.tasks.Task;
 import java.util.Scanner;
 
 /**
