+++ src/main/java/duke/commands/CreateTodoCommand.java
@@ -1,15 +1,11 @@
-package Duke.Commands;
+package duke.commands;
 
-import Duke.DukeException;
-import Duke.Storage;
-import Duke.TaskList;
-import Duke.Tasks.Deadline;
-import Duke.Tasks.Event;
-import Duke.Tasks.Task;
-import Duke.Tasks.Todo;
-import Duke.Ui;
-
-import java.time.DateTimeException;
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.tasks.Task;
+import duke.tasks.Todo;
+import duke.Ui;
 
 /**
  * Represents a command to create a Todo task.
